User disk space properties are now doubles
Very lengthy debugging revealed that using longs would fail to
deserialize with very large numbers due to the fact that Box is
serializing them in scientific notation.  Whoops.  Luckily .Net seems to
know how to convert these to doubles...

+++ BoxApi.V2/Model/ManagedUser.cs
@@ -14,7 +14,7 @@ namespace BoxApi.V2.Model
         ///     The user’s total available space amount in bytes
         /// </summary>
         [JsonProperty(PropertyName = "space_amount")]
-        public long SpaceAmount { get; set; }
+        public double SpaceAmount { get; set; }
 
         /// <summary>
         ///     If the user is part of an enterprise, then this will reflect whether they are a regular user or an admin.
